WIP refactor to export things
