+++ src/client/VZSidebar/styles.scss
