+++ src/index.ts
@@ -1,0 +1,26 @@
+export type {
+  FileId,
+  Files,
+  File,
+  FileTree,
+  FileTreeFile,
+  ShareDBDoc,
+  VZCodeContent,
+} from './types';
+export type { EditorCache } from './client/useEditorCache';
+export { ThemeLabel } from './client/themes';
+export { VZSidebar } from './client/VZSidebar';
+export { VZSettings } from './client/VZSettings';
+export { SplitPaneResizeContext } from './client/SplitPaneResizeContext';
+export { Resizer } from './client/Resizer';
+export { TabList } from './client/TabList';
+export { CodeEditor } from './client/CodeEditor';
+export { PrettierErrorOverlay } from './client/PrettierErrorOverlay';
+export { PresenceNotifications } from './client/PresenceNotifications';
+export { useFileCRUD } from './client/useFileCRUD';
+export { useSubmitOperation } from './client/useSubmitOperation';
+export { vzReducer } from './client/vzReducer';
+export {
+  defaultTheme,
+  useDynamicTheme,
+} from './client/themes';
